Added custom scene graph cloning support. Added multi-step file searching support for shader files' include directive. Modified scene graph structure so that Mesh is not longer a valid scene graph object. It is just a property class owned by Visual. Modified model importing procedure so that importer contains no information about vertex format and vertex buffer binding. Model importing becomes more flexible and more extendable. Modified VisualEffectInstancePass, so that the instancing information is now stored in VisualEffectInstancePass. You could use different instancing number for different passes. Modified PrimitiveTriangles to support triangle strips and fans. Modified the VisualEffect instancing mechanism so that VertexFormat and VertexGroup is provided in VisualEffect's derived classes. Modified Renderer's behavior to allow ignorance toward uninitialized visual effect instance. Replaced through-game-engine accessor method with global thread-safe singleton pattern accessor for object like GameEngineSettings, GameEngineData etc ...

+++ include/FalconEngine/Graphics/Renderer/Resource/VertexBuffer.h
@@ -16,7 +16,7 @@ public:
     /************************************************************************/
     /* Constructors and Destructor                                          */
     /************************************************************************/
-    VertexBuffer(size_t vertexNum, size_t vertexByteNum, BufferUsage usage);
+    VertexBuffer(int vertexNum, size_t vertexByteNum, BufferUsage usage);
     virtual ~VertexBuffer();
 };
