+++ src/FalconEngine/Graphics/Renderer/Resource/VertexBuffer.cpp
@@ -6,7 +6,7 @@ namespace FalconEngine
 /************************************************************************/
 /* Constructors and Destructor                                          */
 /************************************************************************/
-VertexBuffer::VertexBuffer(size_t vertexNum, size_t vertexByteNum, BufferUsage usage) :
+VertexBuffer::VertexBuffer(int vertexNum, size_t vertexByteNum, BufferUsage usage) :
     Buffer(vertexNum, vertexByteNum, usage)
 {
 }
